+++ lib/jx.ex
@@ -33,7 +33,7 @@ defmodule Jx do
       nil ->
         raise ArgumentError
       name ->
-        put_in context.binding[name], value
+        put_in(context.binding[name], value)
     end
   end
 
@@ -93,38 +93,26 @@ defmodule Jx do
     macro_j(expr)
   end
 
-  defp macro_j({:<-, _, [args, context_var]}) do
-    {_left, acc} = parse_left(args, MapSet.new)
-
-    variable_matchers =
-      acc
-      |> Enum.filter(fn :j -> false; _ -> true end)
-      |> Enum.flat_map(fn
-        (:j) -> []
-        (name) ->
-          [{name, Macro.var(name, nil)}]
-      end)
-      |> Enum.sort
-    match_lhs = quote do
-      %Jx{binding: %{unquote_splicing(variable_matchers)}, no_match: false}
-    end
-
-    quote do
-      unquote(match_lhs) = Jx.FunctionMatching.next(unquote(context_var))
-    end
-  end
-
   defp macro_j({:=, meta, [left, right]}) do
-    {left, acc} = parse_left(left, MapSet.new)
-    {right, acc} = parse_right(right, acc)
+    {left, vars_left} = parse_left(left, MapSet.new)
+    {has_j, right, vars_right} = parse_right(right) # XXX: handling has_j=false
+    vars = MapSet.union(vars_left, vars_right)
 
-    term = Macro.escape({:=, meta, [left, right]})
+    right = if has_j do
+      right
+    else
+      quote do
+        Macro.escape(unquote(right))
+      end
+    end
+
+    term = {:{}, [], [:=, meta, [Macro.escape(left), right]]}
     match_rhs = quote do
       Jx.J.j(unquote(term))
     end
 
     variable_matchers =
-      acc
+      vars
       |> Enum.filter(fn :j -> false; _ -> true end)
       |> Enum.flat_map(fn
         (:j) -> []
@@ -148,18 +136,21 @@ defmodule Jx do
   defp parse_left(term, acc) do
     case term do
       {name, _meta, context} when is_atom(name) and is_atom(context) ->
-        # if name |> Atom.to_string |> String.starts_with?("j") do
-        #   raise ArgumentError
-        # end
+        if name |> Atom.to_string |> String.starts_with?("j") do
+          raise ArgumentError
+        end
 
         {term, MapSet.put(acc, name)}
 
       {:^, _, _} -> # Pin operator not supported yet.
         raise ArgumentError
 
-      {name, meta, args} when is_atom(name) and is_list(args) ->
+      {:{}, meta, args} when is_list(args) ->
         {args, acc} = parse_left(args, acc)
-        {{name, meta, args}, acc}
+        {{:{}, meta, args}, acc}
+
+      {_, _, _} ->
+        raise ArgumentError
 
       {a, b} ->
         {[a, b], acc} = Enum.map_reduce([a, b], acc, &parse_left/2)
@@ -173,43 +164,112 @@ defmodule Jx do
     end
   end
 
-  defp parse_right({a, b}, acc) do
-    {a, acc} = parse_right(a, acc)
-    {b, acc} = parse_right(b, acc)
-    {{a, b}, acc}
+  # Tuple literal
+  defp parse_right({a, b}) do
+    {has_j_a, a, vars_a} = parse_right(a)
+    {has_j_b, b, vars_b} = parse_right(b)
+    vars = MapSet.union(vars_a, vars_b)
+
+    case {has_j_a, has_j_b} do
+      {true, true} -> {true, {a, b}, vars}
+      {false, false} -> {false, {a, b}, vars}
+      {true, false} ->
+        b = quote do
+          Macro.escape(unquote(b))
+        end
+        {true, {a, b}, vars}
+      {false, true} ->
+        a = quote do
+          Macro.escape(unquote(a))
+        end
+        {true, {a, b}, vars}
+    end
   end
 
-  defp parse_right(args, acc) when is_list(args) do
-    Enum.map_reduce(args, acc, &parse_right/2)
+  # List term
+  defp parse_right(list) when is_list(list) do
+    {terms, {any_has_j, vars}} = Enum.map_reduce(list, {false, MapSet.new()}, fn term, {has_j0, vars0} ->
+      {has_j, term, vars} = parse_right(term)
+      {{has_j, term}, {has_j0 or has_j, MapSet.union(vars0, vars)}}
+    end)
+
+    terms = if any_has_j do
+      Enum.map(terms, fn 
+        ({true, term}) -> term
+        ({false, term}) ->
+          quote do
+            Macro.escape(unquote(term))
+          end
+      end)
+    else
+      Enum.map(terms, &elem(&1, 1))
+    end
+
+    {any_has_j, terms, vars}
   end
 
-  defp parse_right({{:., dot_meta, dot_args}, meta, args}, acc) when is_list(args) do
-    function_variable_name = case dot_args do
+  # Function calls
+  defp parse_right({{:., dot_meta, dot_args}, meta, args}) when is_list(args) do
+    fn_variable_name = case dot_args do
       [{name, _, _}] when is_atom(name) -> name
       _ -> nil
     end
 
-    is_j_variable = function_variable_name |> Atom.to_string |> String.starts_with?("j")
-
+    is_j_variable = fn_variable_name |> Atom.to_string |> String.starts_with?("j")
     if is_j_variable do
-      meta = Keyword.put(meta, :jx_name, function_variable_name)
-      acc = MapSet.put(acc, function_variable_name)
-      {args, acc} = parse_right(args, acc)
-      {{:j, meta, args}, acc}
+      meta = Keyword.put(meta, :jx_name, fn_variable_name)
+      
+      {has_j, args, vars} = parse_right(args)
+      args = if has_j do
+        args
+      else
+        quote do
+          Macro.escape(unquote(args))
+        end
+      end
+      vars = [fn_variable_name] |> MapSet.new |> MapSet.union(vars)
+      
+      {true, {:{}, [], [:j, meta, args]}, vars}
+    
     else
-      {dot_args, acc} = parse_right(dot_args, acc)
-      {args, acc} = parse_right(args, acc)
+      {has_j, dot_args, _} = parse_right(dot_args) # XXX: generalize to true and handling vars
+      if has_j, do: throw :unimplemented
 
-      {{{:., dot_meta, dot_args}, meta, args}, acc}
+      {has_j, args, vars} = parse_right(args)
+
+      expr = if has_j do
+        {:{}, [], [{:{}, [], [:., dot_meta, dot_args]}, meta, args]}
+      else
+        {{:., dot_meta, dot_args}, meta, args}
+      end 
+
+      {has_j, expr, vars}
     end
   end
 
-  defp parse_right({name, meta, args}, acc) when is_atom(name) and is_list(args) do
-    {args, acc} = parse_right(args, acc)
-    {{name, meta, args}, acc}
+  # Function terms
+  defp parse_right({name, meta, args}) when is_atom(name) and is_list(args) do
+    case parse_right(args) do
+      {false, args, vars} -> {false, {name, meta, args}, vars}
+      
+      {true, args, vars} ->
+        {true, {:{}, [], [name, meta, args]}, vars}
+    end
   end
 
-  defp parse_right(term, acc) do
-    {term, acc}
+  # Variable terms
+  defp parse_right({name, meta, context} = term) when is_atom(name) and is_atom(context) do
+    is_j_variable = name |> Atom.to_string |> String.starts_with?("j")
+    if is_j_variable do
+      meta = put_in(meta[:jx_is_term], true)
+      parse_right({{:., [], [term]}, meta, []})
+    else
+      {false, term, MapSet.new()}
+    end
+  end
+
+  # Literal
+  defp parse_right(term) do
+    {false, term, MapSet.new()}
   end
 end